Move organisation to repo layer
